requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community